+++ api/src/service/domain/workflow/workflowitem_updated.ts
@@ -29,21 +29,10 @@ export interface Modification {
 export const modificationSchema = Joi.object({
   displayName: Joi.string(),
   description: Joi.string().allow(""),
-  exchangeRate: conversionRateSchema
-    .when("amountType", { is: Joi.valid("N/A"), then: Joi.forbidden() }),
-  billingDate: Joi.date()
-    .iso()
-    .when("amountType", { is: Joi.valid("N/A"), then: Joi.forbidden() }),
-  amount: moneyAmountSchema
-    .when("amountType", {
-      is: Joi.valid("N/A"),
-      then: Joi.forbidden(),
-    }),
-  currency: Joi.string()
-    .when("amountType", {
-      is: Joi.valid("N/A"),
-      then: Joi.forbidden(),
-    }),
+  exchangeRate: conversionRateSchema,
+  billingDate: Joi.date().iso(),
+  amount: moneyAmountSchema,
+  currency: Joi.string(),
   amountType: Joi.valid("N/A", "disbursed", "allocated"),
   dueDate: Joi.date().iso(),
   documents: Joi.array().items(storedDocumentSchema),
@@ -186,7 +175,7 @@ function updateDocuments(workflowitem: Workflowitem.Workflowitem, documents?: St
       if (existingDocument.hash !== document.hash) {
         throw new VError(
           `cannot update document ${document.id}, ` +
-          `as changing existing documents is not allowed`,
+            `as changing existing documents is not allowed`,
         );
       }
     }