+++ frontend/src/pages/Workflows/Workflow.js
@@ -48,6 +48,7 @@ const Workflow = props => {
           <ContentAdd />
         </Fab>
         <Fab
+          id="subproject-history-button"
           size="small"
           disabled={props.workflowSortEnabled}
           onClick={() => props.openHistory(props.projectId, props.subProjectId, props.offset, props.limit)}